moved to session recording to correct panels folder
